# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brainiac",
    "description": "GitOps for organizational AI knowledge — capture, govern, and serve an organization's memory. Every read is RLS-scoped to the caller: an agent can never retrieve what its operator cannot.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8600",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/onboard/poll": {
      "post": {
        "tags": [
          "onboard"
        ],
        "summary": "POST /v1/onboard/poll — the CLI's wait loop; the approving poll mints.",
        "description": "Poll a pairing request by device_code. While pending returns {status: \"pending\"}; the first poll after approval claims the request and returns the minted project-scoped key EXACTLY ONCE.",
        "operationId": "onboard_poll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardPollBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Current pairing state (token present exactly once, on claim)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardPollResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown device code"
          }
        }
      }
    },
    "/v1/onboard/requests": {
      "get": {
        "tags": [
          "onboard"
        ],
        "summary": "GET /v1/onboard/requests — the console's approval queue (admin).",
        "description": "Pending pairing requests with their whitelist match in the caller's org (admin). A null project_name means approval will be refused until the remote is registered under a project.",
        "operationId": "requests_list",
        "responses": {
          "200": {
            "description": "Pending requests, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardRequestsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          }
        }
      }
    },
    "/v1/onboard/requests/{id}/approve": {
      "post": {
        "tags": [
          "onboard"
        ],
        "summary": "POST /v1/onboard/requests/{id}/approve — approve into the whitelist match.",
        "description": "Approve a pending pairing (admin). The project is DERIVED from the whitelist — the remote must already be registered under a project in the caller's org, or this refuses with 409.",
        "operationId": "request_approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pairing request id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved; the CLI's next poll mints the key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardDecisionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          },
          "404": {
            "description": "Request not found, expired, or already decided"
          },
          "409": {
            "description": "Remote not whitelisted under any project in this org"
          }
        }
      }
    },
    "/v1/onboard/requests/{id}/deny": {
      "post": {
        "tags": [
          "onboard"
        ],
        "summary": "POST /v1/onboard/requests/{id}/deny — refuse a pending pairing.",
        "description": "Deny a pending pairing request (admin). The polling CLI is told {status: \"denied\"} and stops.",
        "operationId": "request_deny",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pairing request id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardDecisionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          },
          "404": {
            "description": "Request not found or already decided"
          }
        }
      }
    },
...
    "/v1/onboard/start": {
      "post": {
        "tags": [
          "onboard"
        ],
        "summary": "POST /v1/onboard/start — open a pairing request (unauthenticated).",
        "description": "Open an onboarding pairing request for a git remote. Unauthenticated — this is how a developer with no credentials acquires one. Returns a short user_code (matched by eye in the console) and a long device_code to poll with.",
        "operationId": "onboard_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardStartBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pairing opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Remote is not a recognizable git remote URL"
          },
          "429": {
            "description": "Too many pending pairing requests; try again shortly"
          }
        }
      }
    },
...
    "/v1/reviews/promotions/{id}/approve": {
      "post": {
        "tags": [
          "reviews"
        ],
        "description": "Approve a pending promotion: moves the memory to the promotion's target status. Requires a maintainer of the owning team.",
        "operationId": "approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Promotion approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not a maintainer of the owning team"
          },
          "404": {
            "description": "Promotion not found or no longer pending"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OnboardDecisionResponse": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "description": "approved | denied"
          }
        }
      },
      "OnboardPollBody": {
        "type": "object",
        "required": [
          "device_code"
        ],
        "properties": {
          "device_code": {
            "type": "string"
          }
        }
      },
      "OnboardPollResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "org_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "description": "pending | approved | denied | expired | claimed"
          },
          "token": {
            "type": [
              "string",
              "null"
            ],
            "description": "The minted key — present exactly once, on the poll that claims an\napproved request. Never retrievable again."
          }
        }
      },
      "OnboardRequestView": {
        "type": "object",
        "required": [
          "id",
          "user_code",
          "remote",
          "label",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The project this remote would land in (whitelist match in the caller's\norg), or null — meaning approval will 409 until the repo is registered."
          },
          "remote": {
            "type": "string"
          },
          "user_code": {
            "type": "string"
          }
        }
      },
      "OnboardRequestsResponse": {
        "type": "object",
        "required": [
          "requests"
        ],
        "properties": {
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardRequestView"
            }
          }
        }
      },
      "OnboardStartBody": {
        "type": "object",
        "required": [
          "remote"
        ],
        "properties": {
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Who is asking — hostname/username, shown to the approver."
          },
          "remote": {
            "type": "string",
            "description": "The checkout's `git remote get-url origin`, any spelling."
          }
        }
      },
      "OnboardStartResponse": {
        "type": "object",
        "required": [
          "user_code",
          "device_code",
          "remote",
          "verification_url",
          "expires_in_secs",
          "poll_interval_secs"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "Long secret the CLI polls with. Appears only here."
          },
          "expires_in_secs": {
            "type": "integer",
            "format": "int64"
          },
          "poll_interval_secs": {
            "type": "integer",
            "format": "int64"
          },
          "remote": {
            "type": "string",
            "description": "The normalized remote the request was recorded under."
          },
          "user_code": {
            "type": "string",
            "description": "Short code the approver matches by eye in the console."
          },
          "verification_url": {
            "type": "string",
            "description": "Where a human approves this — the console's projects module."
          }
        }
      },
...
      "ReviewDecisionResponse": {
        "type": "object",
        "description": "Outcome of an approve/reject: the promotion, the memory it moved, and the\nstatus that memory now carries.",
        "required": [
          "promotion_id",
          "memory_id",
          "decision",
          "memory_status"
        ],
        "properties": {
          "decision": {
            "type": "string",
            "description": "`approved` | `denied`"
          },
          "memory_id": {
            "type": "string",
            "format": "uuid"
          },
          "memory_status": {
            "type": "string",
            "description": "The memory's status after the decision."
          },
          "promotion_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Env-map token (unrestricted) or a managed `brk_…` key carrying the read/write/admin scope the route requires."
      }
    }
  },
...
}